requirements.txt
feedparser
requests
beautifulsoup4
readability-lxml